+++ src/NW.UnivariateForecasting/Forecasts/ForecastingInitManager.cs
@@ -24,12 +24,12 @@ namespace NW.UnivariateForecasting.Forecasts
 
         #region Methods_public
 
-        public ForecastingInit ExpandValues(ForecastingInit forecastingInit, decimal nextValue)
+        public ForecastingInit ExpandValues(ForecastingInit forecastingInit, double nextValue)
         {
 
             Validator.ValidateObject(forecastingInit, nameof(forecastingInit));
 
-            List<decimal> expandedValues = new List<decimal>();
+            List<double> expandedValues = new List<double>();
             expandedValues.AddRange(forecastingInit.Values);
             expandedValues.Add(nextValue);
